#136-不要なUIとそのファイル、読み込みを削除
